requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
cloudinary==1.33.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.10
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
packaging==25.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.2.0
urllib3==1.26.15
